# apps/avo-api/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AVO API Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "admin@avo.local"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://api.avo.local",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/health": {
            "get": {
                "tags": [
                    "System"
                ],
                "summary": "Health Check",
                "description": "Returns the health status of the API",
                "operationId": "a1dd1f4ddb17ebda7dbe6cb095703f15",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/jobs": {
            "get": {
                "tags": [
                    "Jobs"
                ],
                "summary": "List all jobs",
                "operationId": "1abe28a541df769e888fe7f8c959b61b",
                "responses": {
                    "200": {
                        "description": "List of jobs",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "title": {
                                                "type": "string"
                                            },
                                            "status": {
                                                "type": "string"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Jobs"
                ],
                "summary": "Create a new job",
                "operationId": "83ff691d73cfbaa0dc0000138df2087e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "description",
                                    "status"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Senior AI Engineer"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "We are looking for a..."
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "DRAFT",
                                            "PUBLISHED",
                                            "CLOSED"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Job created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer"
                                        },
                                        "title": {
                                            "type": "string"
                                        },
                                        "status": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/settings": {
            "get": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "List platform settings",
                "operationId": "4ebfe568eb473d8219ca946aa5a5c1d5",
                "responses": {
                    "200": {
                        "description": "Settings"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Create a platform setting",
                "operationId": "edf184b986c91e1cedc4b0963998f2ca",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "key"
                                ],
                                "properties": {
                                    "key": {
                                        "type": "string"
                                    },
                                    "value": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Setting created"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/settings/{key}": {
            "put": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Update a platform setting",
                "operationId": "4761a4f3678e7fd690f91c3a981133c2",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "value": {
                                        "type": "object"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Setting updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin Settings"
                ],
                "summary": "Delete a platform setting",
                "operationId": "727b305b8e6bb37ccafa54f4988e5449",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Setting deleted"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/dashboard/metrics": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Get Admin Dashboard Metrics",
                "operationId": "1e8aeedc00c36ccd873d2074942e85dd",
                "responses": {
                    "200": {
                        "description": "Dashboard Metrics"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/logs": {
            "get": {
                "tags": [
                    "Admin Logs"
                ],
                "summary": "List activity logs",
                "operationId": "1aa74ede5541fd55abbc9e44329a7138",
                "responses": {
                    "200": {
                        "description": "Logs list"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users": {
            "get": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "List users",
                "operationId": "7513a08c2ba82c9a7b6d8b84e7df4624",
                "responses": {
                    "200": {
                        "description": "List of users"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/users/{id}/toggle-status": {
            "patch": {
                "tags": [
                    "Admin Users"
                ],
                "summary": "Toggle user status",
                "operationId": "bbe326c9135c1f182141e800aa11f1d9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Authenticate user and return JWT",
                "operationId": "f5737723279f7c9b81335f452ca38253",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "granix@yopmail.io"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "Admin@1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful login",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "token": {
                                            "type": "string"
                                        },
                                        "user": {
                                            "properties": {
                                                "name": {
                                                    "type": "string"
                                                },
                                                "email": {
                                                    "type": "string"
                                                },
                                                "role": {
                                                    "type": "string"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "expires_in": {
                                            "type": "integer"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "System",
            "description": "System"
        },
        {
            "name": "Jobs",
            "description": "Jobs"
        },
        {
            "name": "Admin Settings",
            "description": "Admin Settings"
        },
        {
            "name": "Admin",
            "description": "Admin"
        },
        {
            "name": "Admin Logs",
            "description": "Admin Logs"
        },
        {
            "name": "Admin Users",
            "description": "Admin Users"
        },
        {
            "name": "Authentication",
            "description": "Authentication"
        }
    ]
}